a lot of work on car ovement, enhanced walljump, camera movement and shield controll. added seat pose for the car
the car movement was reworked to adapt to the mew model with modular wheels and seats. the caracter now abord the car with a coherent pose. the car still need more work to move more realistically. The wall jump now blocks the movement when stick to the wall, this have to be reworked because it couse the caracter to paralize some times.
Now the camera can change of target GameObjct, it stay at a fixed height from the target and it can be rotated with the left stick.
No that much work in the shield, it can be shown and hide now.